+++ runelite-client/src/main/resources/net/runelite/client/plugins/gpu/vert.glsl
@@ -36,14 +36,14 @@ layout(location = 0) in ivec4 VertexPosition;
 layout(location = 1) in vec4 uv;
 
 layout(std140) uniform uniforms {
-  int cameraYaw;
-  int cameraPitch;
+  float cameraYaw;
+  float cameraPitch;
   int centerX;
   int centerY;
   int zoom;
-  int cameraX;
-  int cameraY;
-  int cameraZ;
+  float cameraX;
+  float cameraY;
+  float cameraZ;
   ivec2 sinCosTable[2048];
 };
 
@@ -83,10 +83,10 @@ void main() {
 
   // the client draws one less tile to the north and east than it does to the south
   // and west, so subtract a tiles width from the north and east edges.
-  int fogWest = max(FOG_SCENE_EDGE_MIN, cameraX - drawDistance);
-  int fogEast = min(FOG_SCENE_EDGE_MAX, cameraX + drawDistance - TILE_SIZE);
-  int fogSouth = max(FOG_SCENE_EDGE_MIN, cameraZ - drawDistance);
-  int fogNorth = min(FOG_SCENE_EDGE_MAX, cameraZ + drawDistance - TILE_SIZE);
+  int fogWest = max(FOG_SCENE_EDGE_MIN, int(cameraX) - drawDistance);
+  int fogEast = min(FOG_SCENE_EDGE_MAX, int(cameraX) + drawDistance - TILE_SIZE);
+  int fogSouth = max(FOG_SCENE_EDGE_MIN, int(cameraZ) - drawDistance);
+  int fogNorth = min(FOG_SCENE_EDGE_MAX, int(cameraZ) + drawDistance - TILE_SIZE);
 
   // Calculate distance from the scene edge
   int xDist = min(vertex.x - fogWest, fogEast - vertex.x);
